Added user ID to hash base
Instead of the hash being deterministic based solely on the message contents, we offset the hash to also factor in user ID.   That way if a phrase is common within a community and is said by multiple people, it's less likely to never be learned than if a single spammer says something a lot.   A phrase can still only be learned once, but the function for whether or not a given phrase is learned takes in a user ID parameter.